Fix format and address comments
