+++ sycl/test-e2e/ThreadSanitizer/check_usm.cpp
@@ -21,8 +21,8 @@ int main() {
 #endif
 
   Q.submit([&](sycl::handler &h) {
-     h.parallel_for<class Test>(
-         sycl::nd_range<1>(32, 8), [=](sycl::nd_item<1> item) { array[0]++; });
+     h.parallel_for<class Test>(sycl::nd_range<1>(32, 8),
+                                [=](sycl::nd_item<1>) { array[0]++; });
    }).wait();
   // CHECK: WARNING: DeviceSanitizer: data race
   // CHECK-NEXT: When write of size 1 at 0x{{.*}} in kernel <{{.*}}Test>
