+++ unified-runtime/source/loader/layers/sanitizer/tsan/tsan_libdevice.hpp
@@ -53,52 +53,6 @@ struct VectorClock {
   Epoch clk_[kThreadSlotCount] = {};
 };
 
-struct Shadow {
-public:
-  static constexpr RawShadow kEmpty = static_cast<RawShadow>(0);
-  
-  static constexpr RawShadow kRodata = static_cast<RawShadow>(1 << 30);
-
-  Shadow(uint32_t addr, uint32_t size, uint32_t clock, uint32_t sid,
-         AccessType typ) {
-    raw_ = 0;
-    raw_ |= (!!(typ & kAccessAtomic) << 31) | (!!(typ & kAccessRead) << 30) |
-            (clock << 16) | (sid << 8) |
-            (((((1u << size) - 1) << (addr & 0x7)) & 0xff));
-  }
-
-  explicit Shadow(RawShadow x = Shadow::kEmpty) {
-    raw_ = static_cast<uint32_t>(x);
-  }
-
-  RawShadow raw() const { return static_cast<RawShadow>(raw_); }
-
-  Sid sid() const { return part_.sid_; }
-
-  uint16_t clock() const { return part_.clock_; }
-
-  uint8_t access() const { return part_.access_; }
-
-  bool IsBothReads(AccessType Type) {
-    uint32_t is_read = !!(Type & kAccessRead);
-    bool res = raw_ & (is_read << 30);
-    return res;
-  }
-
-private:
-  struct Parts {
-    uint8_t access_;
-    Sid sid_;
-    uint16_t clock_ : 14;
-    uint16_t is_read_ : 1;
-    uint16_t is_atomic_ : 1;
-  };
-  union {
-    Parts part_;
-    uint32_t raw_;
-  };
-};
-
 struct TsanErrorReport {
   int Flag = 0;
 
@@ -132,9 +86,9 @@ struct TsanRuntimeData {
   DeviceType DeviceTy = DeviceType::UNKNOWN;
 
   uint32_t Debug = 0;
-  
+
   int Lock = 0;
-  
+
   uint32_t RecordedReportCount = 0;
 
   TsanErrorReport Report[TSAN_MAX_NUM_REPORTS];
