+++ unified-runtime/source/loader/layers/sanitizer/tsan/tsan_shadow.hpp
@@ -54,18 +54,20 @@ struct ShadowMemory {
   uptr ShadowEnd = 0;
 };
 
-/// Shadow Memory layout of CPU device
-///
-/// 0000 0000 1000 - 0200 0000 0000: main binary and/or MAP_32BIT mappings (2TB)
-/// 0200 0000 0000 - 1000 0000 0000: -
-/// 1000 0000 0000 - 3000 0000 0000: shadow (32TB)
-/// 3000 0000 0000 - 3800 0000 0000: metainfo (memory blocks and sync objects; 8TB)
-/// 3800 0000 0000 - 5500 0000 0000: -
-/// 5500 0000 0000 - 5a00 0000 0000: pie binaries without ASLR or on 4.1+ kernels
-/// 5a00 0000 0000 - 7200 0000 0000: -
-/// 7200 0000 0000 - 7300 0000 0000: heap (1TB)
-/// 7300 0000 0000 - 7a00 0000 0000: -
-/// 7a00 0000 0000 - 8000 0000 0000: modules and main thread stack (6TB)
+// clang-format off
+// Shadow Memory layout of CPU device
+//
+// 0000 0000 1000 - 0200 0000 0000: main binary and/or MAP_32BIT mappings (2TB)
+// 0200 0000 0000 - 1000 0000 0000: -
+// 1000 0000 0000 - 3000 0000 0000: shadow (32TB)
+// 3000 0000 0000 - 3800 0000 0000: metainfo (memory blocks and sync objects; 8TB)
+// 3800 0000 0000 - 5500 0000 0000: -
+// 5500 0000 0000 - 5a00 0000 0000: pie binaries without ASLR or on 4.1+ kernels
+// 5a00 0000 0000 - 7200 0000 0000: -
+// 7200 0000 0000 - 7300 0000 0000: heap (1TB)
+// 7300 0000 0000 - 7a00 0000 0000: -
+// 7a00 0000 0000 - 8000 0000 0000: modules and main thread stack (6TB)
+// clang-format on
 struct ShadowMemoryCPU final : public ShadowMemory {
   ShadowMemoryCPU(ur_context_handle_t Context, ur_device_handle_t Device)
       : ShadowMemory(Context, Device) {}
@@ -100,13 +102,15 @@ struct ShadowMemoryGPU : public ShadowMemory {
   std::unordered_map<uptr, ur_physical_mem_handle_t> VirtualMemMaps;
 };
 
-/// Shadow Memory layout of GPU PVC device
-/// We only support limited memory range for host/shared usm
-/// USM Allocation Range (56 bits)
-///   Host   USM : 0x00ff_ff00_0000_0000 ~ 0x00ff_ffff_ffff_ffff
-///   Shared USM : 0x0000_7f00_0000_0000 ~ 0x0000_7fff_ffff_ffff
-///   Device USM : 0xff00_0000_0000_0000 ~ 0xff00_ffff_ffff_ffff
-///
+// clang-format off
+// Shadow Memory layout of GPU PVC device
+// We only support limited memory range for host/shared usm
+// USM Allocation Range (56 bits)
+//   Host   USM : 0x00ff_ff00_0000_0000 ~ 0x00ff_ffff_ffff_ffff
+//   Shared USM : 0x0000_7f00_0000_0000 ~ 0x0000_7fff_ffff_ffff
+//   Device USM : 0xff00_0000_0000_0000 ~ 0xff00_ffff_ffff_ffff
+//
+// clang-format on
 struct ShadowMemoryPVC : public ShadowMemoryGPU {
   ShadowMemoryPVC(ur_context_handle_t Context, ur_device_handle_t Device)
       : ShadowMemoryGPU(Context, Device) {}
